==> 4) Predict Winner with Logistic Regression.ipynb <==
# 4) Predicting Winners with Logistic Regression

Finding a linear model to predict how many points a team will score had too many variables to accurately predict the winner of a game. Instead, we can try to use a logistic regression to see if a team will win or lose.

#Imports
import pandas as pd
pd.set_option('display.max_columns', None)

import seaborn as sns
import numpy as np

from sklearn.linear_model import LogisticRegression
from sklearn import metrics
from sklearn.model_selection import train_test_split

#Read in the data
df = pd.read_csv(r'nfl_game_logs_df.csv', index_col=0)

The null model would select a team at random and predict them to win. This would mean the picks are accurate just under 50% of the time (assuming even odds a team will win or lose, and that ties are infrequent).

We can make other predictions with minimal effort that produce better results. For instance, the home team wins about 55% of the time compared to the away team.

Additionally, the Elo scoring system I implented shows that the team with the higher score before the game wins about 63.5% of the time.

df[df['home_team']]['result'].value_counts(normalize=True).to_frame()

df.query('elo_start > elo_start_opp')['result'].value_counts(normalize=True).to_frame()

For simplicity, we can predict team wins instead of ties or losses.

df['result_letter'] = df['result_value'].map({1.0: 'W', 0.5: 'LT', 0.0: 'LT'})

logreg = LogisticRegression()

feature_cols = ['elo_opp_diff_team',
                '8_game_avg_points_for',
                '8_game_avg_points_for_opp',
                '4_game_avg_exp_pts_off',
                'home_team'
               ]

X = df[feature_cols]
y = df['result_letter']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=123)

LR = LogisticRegression()
LR.fit(X_train, y_train)

pred = LR.predict(X_test)

LR.score(X, y)

LR.score(X_train, y_train)

LR.score(X_test, y_test)
